fix: optimize loop counter types in slice functions (#159)

+++ contracts/ethereum/contracts/AjoCircle.sol
@@ -113,7 +113,7 @@ contract AjoCircle is Initializable, OwnableUpgradeable, ReentrancyGuardUpgradea
         }
 
         address[] memory slice = new address[](count);
-        for (uint32 i = 0; i < count; i++) {
+        for (uint256 i = 0; i < count; i++) {
             slice[i] = memberAddresses[_offset + i];
         }
         return slice;
@@ -127,7 +127,7 @@ contract AjoCircle is Initializable, OwnableUpgradeable, ReentrancyGuardUpgradea
         }
 
         address[] memory slice = new address[](count);
-        for (uint32 i = 0; i < count; i++) {
+        for (uint256 i = 0; i < count; i++) {
             slice[i] = rotationOrder[_offset + i];
         }
         return slice;
